requirements.txt
moviepy
praw
pysqlite3
requests
simple_youtube_api
schedule
pytiktokapi
instagrapi

